Redis implementation of the session data store.
