+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/cache/AuthenticationRequestCache.java
@@ -35,8 +35,8 @@ import org.wso2.carbon.identity.core.util.IdentityUtil;
 public class AuthenticationRequestCache extends
         BaseCache<AuthenticationRequestCacheKey, AuthenticationRequestCacheEntry> {
 
-    private static Log log = LogFactory.getLog(AuthenticationRequestCache.class);
     private static final String AUTHENTICATION_REQUEST_CACHE_NAME = "AuthenticationRequestCache";
+    private static Log log = LogFactory.getLog(AuthenticationRequestCache.class);
     private static volatile AuthenticationRequestCache instance;
     private boolean isTemporarySessionDataPersistEnabled = false;
 
@@ -44,6 +44,7 @@ public class AuthenticationRequestCache extends
      * Private constructor which will not allow to create objects of this class from outside
      */
     private AuthenticationRequestCache() {
+
         super(AUTHENTICATION_REQUEST_CACHE_NAME, true);
         if (IdentityUtil.getProperty("JDBCPersistenceManager.SessionDataPersist.Temporary") != null) {
             isTemporarySessionDataPersistEnabled = Boolean.parseBoolean(
@@ -57,6 +58,7 @@ public class AuthenticationRequestCache extends
      * @return AuthenticationRequestCache
      */
     public static AuthenticationRequestCache getInstance() {
+
         if (instance == null) {
             synchronized (AuthenticationRequestCache.class) {
                 if (instance == null) {
@@ -74,7 +76,8 @@ public class AuthenticationRequestCache extends
      * @param entry Actual object where cache entry is placed.
      */
     public void addToCache(AuthenticationRequestCacheKey key, AuthenticationRequestCacheEntry entry) {
-        super.addToCache(key,entry);
+
+        super.addToCache(key, entry);
         if (isTemporarySessionDataPersistEnabled) {
             int tenantId = MultitenantConstants.INVALID_TENANT_ID;
             String tenantDomain = entry.getAuthenticationRequest().getTenantDomain();
@@ -93,6 +96,7 @@ public class AuthenticationRequestCache extends
      * @return Cached entry.
      */
     public AuthenticationRequestCacheEntry getValueFromCache(AuthenticationRequestCacheKey key) {
+
         AuthenticationRequestCacheEntry entry = super.getValueFromCache(key);
         if (entry == null && isTemporarySessionDataPersistEnabled) {
             entry = (AuthenticationRequestCacheEntry) SessionDataStore.getInstance().
@@ -106,7 +110,8 @@ public class AuthenticationRequestCache extends
      *
      * @param key Key to clear cache.
      */
-    public void clearCacheEntry(AuthenticationRequestCacheKey key){
+    public void clearCacheEntry(AuthenticationRequestCacheKey key) {
+
         super.clearCacheEntry(key);
         if (isTemporarySessionDataPersistEnabled) {
             SessionDataStore.getInstance().clearSessionData(key.getResultId(), AUTHENTICATION_REQUEST_CACHE_NAME);
