+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/store/SessionDataPersistTask.java
@@ -24,7 +24,7 @@ import org.apache.commons.logging.LogFactory;
 import java.util.concurrent.BlockingDeque;
 
 /**
- * Task to persist and remove session data
+ * Task to persist and remove session data.
  */
 public class SessionDataPersistTask implements Runnable {
